requirements.txt
schedule==1.2.2
python-dotenv==1.1.0
